add ArchiveModule.ts
