--- vscode-plugin/synapse-schemas/data_service.xsd
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<!--
-  ~  Licensed to the Apache Software Foundation (ASF) under one
-  ~  or more contributor license agreements.  See the NOTICE file
-  ~  distributed with this work for additional information
-  ~  regarding copyright ownership.  The ASF licenses this file
-  ~  to you under the Apache License, Version 2.0 (the
-  ~  "License"); you may not use this file except in compliance
-  ~  with the License.  You may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~  Unless required by applicable law or agreed to in writing,
-  ~  software distributed under the License is distributed on an
-  ~   * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~  KIND, either express or implied.  See the License for the
-  ~  specific language governing permissions and limitations
-  ~  under the License.
-  -->
-
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" >
-
-    <xs:element name="data" type="DATA">
-        <xs:annotation>
-            <xs:documentation source="description">
-                Data defines configurations for creating data services
-            </xs:documentation>
-        </xs:annotation>
-    </xs:element>
-
-    <xs:complexType name="DATA" >
-        <!-- TODO : need to add the constraints (might need factory refactoring) -->
-        <xs:annotation>
-            <xs:documentation source="description">
-                This describes the possible configurations for data services
-            </xs:documentation>
-        </xs:annotation>
-        <xs:sequence>
-            <xs:element name="config" minOccurs="0" maxOccurs="unbounded">
-            </xs:element>
-            <xs:element name="query" minOccurs="0" maxOccurs="unbounded">
-            </xs:element>
-            <xs:element name="operation" minOccurs="0" maxOccurs="unbounded">
-            </xs:element>
-            <xs:element name="resource" minOccurs="0" maxOccurs="unbounded">
-            </xs:element>
-            <xs:element name="event-trigger" minOccurs="0" maxOccurs="unbounded">
-            </xs:element>
-        </xs:sequence>
-
-        <xs:attribute name="baseURI" type="xs:anyURI" use="required"/>
-        <xs:attribute name="name" type="xs:NMTOKEN" use="required"/>
-
-        <xs:attribute name="enableBatchRequests" type="xs:boolean" use="optional"/>
-        <xs:attribute name="enableBoxcarring" type="xs:boolean" use="optional"/>
-        <xs:attribute name="txManagerJNDIName" type="xs:NMTOKEN" use="optional"/>
-        <xs:attribute name="serviceNamespace" type="xs:anyURI" use="optional"/>
-        <xs:attribute name="serviceGroup" type="xs:NMTOKEN" use="optional"/>
-        <xs:attribute name="serviceStatus" use="optional">
-            <xs:simpleType>
-                <xs:restriction base="xs:string">
-                    <xs:enumeration value="active" />
-                    <xs:enumeration value="inactive" />
-                </xs:restriction>
-            </xs:simpleType>
-        </xs:attribute>
-        <xs:attribute name="transports" type="xs:string" use="optional"/>
-    </xs:complexType>
-
-</xs:schema>
