+++ vscode-plugin/synapse-schemas/synapse_config.xsd
@@ -33,9 +33,6 @@
     <xs:include schemaLocation="task.xsd" />
     <xs:include schemaLocation="template.xsd" />
     <xs:include schemaLocation="registry.xsd" />
-    <xs:include schemaLocation="data_service.xsd" />
-
-
 
     <xs:element name="definitions" type="Definition">
         <xs:annotation>
@@ -66,7 +63,6 @@
             <xs:element ref="sequence" minOccurs="0" maxOccurs="unbounded" />
             <xs:element ref="task" minOccurs="0" maxOccurs="unbounded" />
             <xs:element ref="template" minOccurs="0" maxOccurs="unbounded" />
-            <xs:element ref="data" minOccurs="0" maxOccurs="unbounded" />
         </xs:choice>
     </xs:complexType>
 </xs:schema>
